pyproject.toml
[project]
name = "drf-p33"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-filter>=25.1",
    "drf-spectacular>=0.28.0",
    "faker>=37.6.0",
    "psycopg2-binary>=2.9.10",
]
